# src/main/resources/config-openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Invoice and Payments Microservice API
  description: Microservicio para gestión de facturas y pagos
  version: 1.0.0
servers:
  - url: http://localhost:8082
    description: Servicio de Facturas y Pagos
paths:
...
  /payment-methods:
    get:
      tags:
        - PaymentMethod
      summary: Obtener todos los métodos de pago
      operationId: getPaymentMethods
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Métodos de pago obtenidos exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentMethodResult'
        '500':
          description: Error interno del servidor

    post:
      tags:
        - PaymentMethod
      summary: Crear un nuevo método de pago
      operationId: createPaymentMethod
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentMethodRequestDTO'
      responses:
        '201':
          description: Método de pago creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentMethodResponseDTO'
        '400':
          description: Solicitud inválida
        '500':
          description: Error interno del servidor

  # ==================== PAYMENT DETAILS ====================
  /payment-details:
    get:
      tags:
        - PaymentDetail
      summary: Obtener todos los detalles de pago
      operationId: getPaymentDetails
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Detalles de pago obtenidos exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentDetailResult'
        '500':
          description: Error interno del servidor

    post:
      tags:
        - PaymentDetail
      summary: Crear un nuevo detalle de pago
      operationId: createPaymentDetail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentDetailRequestDTO'
      responses:
        '201':
          description: Detalle de pago creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentDetailResponseDTO'
        '400':
          description: Solicitud inválida
        '500':
          description: Error interno del servidor
components:
  schemas:
...
    PaymentMethodRequestDTO:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 60
    PaymentMethodResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    PaymentMethodResult:
      type: object
      properties:
        totalCount:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/PaymentMethodResponseDTO'

    # ==================== PAYMENT DETAIL ====================
    PaymentDetailRequestDTO:
      type: object
      required:
        - invoiceId
        - paymentMethodId
        - appliedAmount
      properties:
        invoiceId:
          type: integer
          format: int64
        paymentMethodId:
          type: integer
          format: int64
        appliedAmount:
          type: number
          format: double
    PaymentDetailResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        active:
          type: boolean
        invoiceId:
          type: integer
          format: int64
        paymentMethodId:
          type: integer
          format: int64
        appliedAmount:
          type: number
          format: double
    PaymentDetailResult:
      type: object
      properties:
        totalCount:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/PaymentDetailResponseDTO'